requirements.txt
numpy>=1.19.2
pandas==1.3.1
transformers==4.9.2
tqdm==4.48.2
torch==1.9.0
tensorflow==2.7.0
scikit-learn==0.24.2
seqeval==1.2.2
pm4py==2.2.16
nltk==3.6.2
spacy==3.1.1
pytorch-transformers==1.2.0
seaborn==0.11.0
matplotlib==3.4.2
gensim==4.1.2
ocel-standard
keras
simpletransformers==0.63.0

